LayoutTests/platform/chromium/http/tests/security/dataURL/xss-DENIED-to-data-url-window-open-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL data:text/html,<html><head><script>function fireSentinel() {if (window.layoutTestController)layoutTestController.globalFlag = true;}</script></head><body onload="fireSentinel();"><p>Opened Frame</p><p id='accessMe'>PASS: Cross frame access from an opener frame was denied</p></body></html> from frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-to-data-url-window-open.html. Domains, protocols and ports must match.

Opener Frame

Test
